Adding keepalives to the bitten client/server protocol.
Keepalives now keep builds running as long as the slave is up, rather than
having them time out independently of if the slave is working on them.

This requires an update of both the master and slave to support keepalives--
older slaves (protocol version 3) will continue to work, but will not submit
keepalives (and thus will retain previous behavior.)

I've updated the PROTOCOL_VERSION field as older masters would kill builds
for newer slaves that attempted to keepalive. We probably want to update
test_create_build_protocol_wrong_version to test the matrix of
allowed versions, but I expect the exception of allowing of older slaves
to disappear when we next edit the protocol.

Closes #411.