#661: unable to inform gcov of object file locations
Assuming the following build tree:

folder/src/test.c folder/obj/test.o

there doesn't seem to be a way to make gcov find the object files. The
root argument is not helpful because the object files are already in the
same root folder as the sources. The prefix argument would be useful as
a way to specify a relative path from the source to the object file.
However, it always includes a hyphen, making it useless for this
purpose.

This problem was earlier reported in #231 (from which came the root
argument).

I've attached a patch that adds a new argument with the specific use of
allowing one to specify a relative path from source to object file.